Moved the PhysiCell_Settings and PhysiCell_Pugixml to the physicellcore folder to handle issues with compilation
