+++ src/modules/CMakeLists.txt
@@ -5,8 +5,6 @@ PhysiCell_geometry.cpp
 PhysiCell_MultiCellDS.cpp
 PhysiCell_pathology.cpp
 PhysiCell_POV.cpp
-PhysiCell_pugixml.cpp
-PhysiCell_settings.cpp
 #PhysiCell_standard_modules.h
 PhysiCell_SVG.cpp
 PhysiCell_various_outputs.cpp
