+++ src/modules/PhysiCell_various_outputs.cpp
@@ -68,7 +68,7 @@
 #include "../physicellcore/PhysiCell.h"
 #include "./PhysiCell_various_outputs.h"
 
-#include "./PhysiCell_settings.h"
+#include "../physicellcore/PhysiCell_settings.h"
 
 namespace PhysiCell{
