+++ src/physicellcore/PhysiCell.h
@@ -84,6 +84,8 @@ static std::string PhysiCell_DOI = "10.1371/journal.pcbi.1005991";
 // #include "PhysiCell_digital_cell_line.h" // to be deprecated! 
 #include "PhysiCell_utilities.h"
 #include "PhysiCell_constants.h"
+#include "PhysiCell_pugixml.h"
+#include "PhysiCell_settings.h"
