+++ src/physicellcore/PhysiCell_phenotype.h
@@ -75,7 +75,7 @@
 
 #include "../BioFVM/BioFVM.h" 
 
-#include "../modules/PhysiCell_settings.h"
+#include "./PhysiCell_settings.h"
 
 using namespace BioFVM;
