+++ src/physicellcore/PhysiCell_settings.h
@@ -80,8 +80,8 @@
 #include "./PhysiCell_pugixml.h"
 #include "../BioFVM/BioFVM.h"
 
-#include "../physicellcore/PhysiCell_constants.h" 
-#include "../physicellcore/PhysiCell_utilities.h"
+#include "./PhysiCell_constants.h" 
+#include "./PhysiCell_utilities.h"
 
 using namespace BioFVM;
